+++ src/components/sets/index.js
@@ -1,2 +1,4 @@
 export { default as CellSetsManagerSubscriber } from './CellSetsManagerSubscriber';
 export { default as SetsManager } from './SetsManager';
+export { default as CellSetSizePlot } from './CellSetSizePlot';
+export { default as CellSetSizePlotSubscriber } from './CellSetSizePlotSubscriber';
