+++ src/components/sets/reducer.test.js
@@ -2,6 +2,7 @@
 import expect from 'expect';
 import reducer, {
   treeInitialize, nodeToRenderProps, ACTION,
+  treeToVisibleSetSizes,
 } from './reducer';
 
 import {
@@ -421,6 +422,21 @@ describe('Hierarchical sets reducer', () => {
       });
       expect(postDragNode.tree[0].children[0].children[0]._state.key).toEqual('vasculature-endothelial');
       expect(postDragNode.tree[0].children[0].children[1]._state.key).toEqual('vasculature-pericytes');
+    });
+
+    it('can convert a tree to an array with visible set sizes', () => {
+      const initialTree = tree;
+      expect(initialTree._state.visibleKeys).toEqual([]);
+      const postViewTree = reducer(initialTree, {
+        type: ACTION.VIEW_NODE,
+        targetKey: 'vasculature-endothelial',
+      });
+      const setSizes = treeToVisibleSetSizes(postViewTree);
+      expect(setSizes.length).toEqual(1);
+      expect(setSizes[0].key).toEqual('vasculature-endothelial');
+      expect(setSizes[0].name).toEqual('Endothelial');
+      expect(setSizes[0].size).toEqual(3);
+      expect(setSizes[0].color).toEqual([100, 0, 0]);
     });
   });
 });
