+++ src/components/vega/index.js
@@ -1,3 +1,2 @@
 export { default as VegaPlot } from './VegaPlot';
-export { default as CellSetSizePlot } from './CellSetSizePlot';
-export { default as CellSetSizePlotSubscriber } from './CellSetSizePlotSubscriber';
+export { createVegaLiteApi, VEGA_THEMES } from './utils';
