+++ src/components/vega/utils.js
@@ -1,39 +1,6 @@
-import { useRef, useState, useEffect } from 'react';
-import PubSub from 'pubsub-js';
 import * as vl from 'vega-lite-api';
 import * as Vega from 'vega';
 import * as VegaLite from 'vega-lite';
-import { GRID_RESIZE } from '../../events';
-
-/**
- * Custom hook, subscribes to GRID_RESIZE and window resize events.
- * @returns {array} `[width, height, containerRef]` where width and height
- * are numbers and containerRef is a React ref.
- */
-export function useGridItemSize() {
-  const containerRef = useRef();
-
-  const [height, setHeight] = useState();
-  const [width, setWidth] = useState();
-
-  useEffect(() => {
-    function onResize() {
-      if (!containerRef.current) return;
-      const containerRect = containerRef.current.getBoundingClientRect();
-      setHeight(containerRect.height);
-      setWidth(containerRect.width);
-    }
-    const gridResizeToken = PubSub.subscribe(GRID_RESIZE, onResize);
-    window.addEventListener('resize', onResize);
-    onResize();
-    return () => {
-      PubSub.unsubscribe(gridResizeToken);
-      window.removeEventListener('resize', onResize);
-    };
-  }, []);
-
-  return [width, height, containerRef];
-}
 
 /**
  * Vega-Lite themes that can be passed to the `config` property
@@ -44,9 +11,7 @@ export const VEGA_THEMES = {
     // The vega-themes dark theme.
     // Reference: https://github.com/vega/vega-themes/blob/master/src/theme-dark.ts
     background: null,
-
     title: { color: '#fff' },
-
     style: {
       'guide-label': {
         fill: '#fff',
@@ -55,7 +20,6 @@ export const VEGA_THEMES = {
         fill: '#fff',
       },
     },
-
     axis: {
       domainColor: '#fff',
       gridColor: '#888',
